Changed vector of requests to be mutable.
Changed the `batchedRequest` vector inside the `Master::Http`
class to be mutable, in order to be able to restore the
`const` qualifier to the `/state`-handler.

While this may be a debatable benefit for this
endpoint alone, the subsequent commits introduce
additional work to extend this mechanism to all
read-only endpoint of `Master`.

Without this change, this would create a paradoxical
situation where all mutating endpoints are marked as
const and all read-only endpoints are marked as
non-const.

Review: https://reviews.apache.org/r/68378