Override source address for executors.
Ignore an executor's reported IP address and enforce the hostname
"localhost" instead.

This is intended to enable TLS connections to executors running
in bridge-mode or virtual-mode docker containers, which can not
present a valid certificate containing their source IP which is
dynamically generated and not known until the executor connects.

NOTE: This change is for illustration purposes only, and is not
intended to be upstreamed.

Review: https://reviews.apache.org/r/70886